Fixed runner problem

+++ backend/problems/two-sum-count-pairs/starter/cpp.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int countPairs(vector<int>& nums, int target) {
+    long long countPairs(vector<int>& nums, int target) {
         return 0;
     }
 };
