+++ backend/problems/two-sum-count-pairs/starter/java.java
@@ -1,5 +1,5 @@
 class Solution {
-    public int countPairs(int[] nums, int target) {
+    public long countPairs(int[] nums, int target) {
         return 0;
     }
 }
